Added main new plugin for using only public api for modules and refactored code for this
