+++ .eslintrc.js
@@ -51,6 +51,7 @@ module.exports = {
         '@typescript-eslint/explicit-function-return-type': 'warn',
         'react/prop-types': 'off',
         'heikun-eslint-checker/path-checker': ['error', {alias: '@'}],
+        'heikun-eslint-checker/public-api-imports': ['error', {alias: '@'}],
         'i18next/no-literal-string': [
             'error',
             {markupOnly: true}
