+++ json-server/db.json
@@ -790,6 +790,12 @@
       "articleId": "17",
       "rate": 2,
       "id": "URcBBXm"
+    },
+    {
+      "userId": "1",
+      "articleId": "7",
+      "rate": 5,
+      "id": "DzCRWIp"
     }
   ],
   "profile-ratings": [
