+++ package.json
@@ -68,7 +68,7 @@
     "css-loader": "^6.8.1",
     "eslint": "^8.53.0",
     "eslint-config-standard-with-typescript": "^39.1.1",
-    "eslint-plugin-heikun-eslint-checker": "^0.0.7",
+    "eslint-plugin-heikun-eslint-checker": "^0.0.9",
     "eslint-plugin-i18next": "^6.0.3",
     "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-n": "^16.3.1",
