+++ src/app/providers/router/ui/AppRouter.tsx
@@ -1,6 +1,6 @@
 import React, { Suspense, memo, type FC, useCallback } from 'react'
 import { Route, Routes } from 'react-router-dom'
-import { PageLoader } from '@/widgets/PageLoader/PageLoader'
+import { PageLoader } from '@/widgets/PageLoader'
 import { type AppRouteProps, routeConfig } from '@/shared/config/routeConfig/routeConfig'
 import { RequireAuth } from '@/app/providers/router/ui/requireAuth'
