+++ src/pages/AboutPage/ui/AboutPage.tsx
@@ -1,6 +1,6 @@
 import { type ReactElement } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Page } from '@/widgets/Page/Page'
+import { Page } from '@/widgets/Page'
 
 const AboutPage = (): ReactElement => {
     const { t } = useTranslation('about')
