+++ src/pages/ArticleDetailsPage/ui/ArticleDetailsPage/ArticleDetailsPage.tsx
@@ -3,7 +3,7 @@ import { classNames } from '@/shared/lib/classNames/classNames'
 import { ArticleDetails } from '@/ourEntities/Article'
 import { useParams } from 'react-router-dom'
 import { DynamicModuleLoader, type ReducersList } from '@/shared/lib/components/DynamicModuleLoader/DynamicModuleLoader'
-import { Page } from '@/widgets/Page/Page'
+import { Page } from '@/widgets/Page'
 import { articleDetailsPageReducer } from '../../model/slices'
 import { ArticleDetailsPageHeader } from '../ArticleDetailsPageHeader/ArticleDetailsPageHeader'
 import { VStack } from '@/shared/ui/Stack'
