+++ src/pages/ArticleEditPage/ui/ArticleEditPage/ArticleEditPage.tsx
@@ -3,7 +3,7 @@ import { classNames } from '@/shared/lib/classNames/classNames'
 import { useTranslation } from 'react-i18next'
 import cls from './ArticleEditPage.module.scss'
 import { useParams } from 'react-router-dom'
-import { Page } from '@/widgets/Page/Page'
+import { Page } from '@/widgets/Page'
 
 interface ArticleEditPageProps {
     className?: string
