+++ src/pages/ArticlesPage/ui/ArticlesPage/ArticlesPage.tsx
@@ -4,7 +4,7 @@ import cls from './ArticlesPage.module.scss'
 import { DynamicModuleLoader, type ReducersList } from '@/shared/lib/components/DynamicModuleLoader/DynamicModuleLoader'
 import { articlesPageReducer } from '../../model/slices/articlePageSlice'
 import { useAppDispatch } from '@/shared/lib/hooks/useAppDispatch/useAppDispatch'
-import { Page } from '@/widgets/Page/Page'
+import { Page } from '@/widgets/Page'
 import { fetchNextArticlesPage } from '../../model/services/fetchNextArticlesPage/fetchNextArticlesPage'
 import { ArticleInfiniteList } from '../ArticleInfiniteList/ArticleInfiniteList'
 import { ArticlesPageFilter } from '../ArticlesPageFilter/ArticlesPageFilter'
