+++ src/pages/ForbiddenPage/ui/ForbiddenPage/ForbiddenPage.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next'
 import { type ReactElement } from 'react'
-import { Page } from '@/widgets/Page/Page'
+import { Page } from '@/widgets/Page'
 import { Text } from '@/shared/ui/Text/Text'
 
 const ForbiddenPage = (): ReactElement => {
