+++ src/pages/MainPage/ui/MainPage.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next'
 import { type ReactElement } from 'react'
-import { Page } from '@/widgets/Page/Page'
+import { Page } from '@/widgets/Page'
 
 const MainPage = (): ReactElement => {
     const { t } = useTranslation('main')
