+++ src/pages/ProfilePage/ui/ProfilePage.tsx
@@ -1,7 +1,7 @@
 import { type FC } from 'react'
 import { classNames } from '@/shared/lib/classNames/classNames'
 import cls from './ProfilePage.module.scss'
-import { Page } from '@/widgets/Page/Page'
+import { Page } from '@/widgets/Page'
 import { VStack } from '@/shared/ui/Stack/VStack/VStack'
 import { EditableProfileCard } from '@/features/editableProfileCard'
 import { useParams } from 'react-router-dom'
