+++ src/widgets/LangSwitcher/index.ts
@@ -1,0 +1,1 @@
+export { LangSwitcher } from './ui/LangSwitcher/LangSwitcher'
