+++ src/widgets/Page/index.ts
@@ -1,0 +1,1 @@
+export { Page, PAGE_ID } from './ui/Page/Page'
