+++ src/widgets/Page/ui/Page/Page.module.scss
