+++ src/widgets/PageLoader/index.ts
@@ -1,0 +1,1 @@
+export { PageLoader } from './ui/PageLoader/PageLoader'
