+++ src/widgets/PageLoader/ui/PageLoader/PageLoader.module.scss
