+++ src/widgets/Sidebar/ui/Sidebar/Sidebar.tsx
@@ -2,7 +2,7 @@ import { type FC, memo, useMemo, useState } from 'react'
 import { classNames } from '@/shared/lib/classNames/classNames'
 import cls from './Sidebar.module.scss'
 import { ThemeSwitcher } from '@/widgets/ThemeSwitcher'
-import { LangSwitcher } from '@/widgets/LangSwitcher/LangSwitcher'
+import { LangSwitcher } from '@/widgets/LangSwitcher'
 import { Button, ButtonSize, ButtonTheme } from '@/shared/ui/Button/Button'
 import { SidebarItem } from '../SidebarItem/SidebarItem'
 import { type SidebarItemType } from '../../model/types/sidebar'
